+++ package.json
@@ -20,7 +20,7 @@
     "./format": "./format/index.js",
     "./guard": "./guard/index.js",
     "./hash": "./hash/index.js",
-    "./settings": "./settings/index.js",
+    "./system": "./system/index.js",
     "./value": "./value/index.js",
     ".": "./typebox.js"
   },
