+++ tsconfig.json
@@ -28,8 +28,8 @@
       "@sinclair/typebox/hash": [
         "src/hash/index.ts"
       ],
-      "@sinclair/typebox/settings": [
-        "src/settings/index.ts"
+      "@sinclair/typebox/system": [
+        "src/system/index.ts"
       ],
       "@sinclair/typebox/value": [
         "src/value/index.ts"
